Renamed user methods file and added stub for tests
